+++ packages/uno-serverless-azure/src/services/documentdb-query.ts
@@ -1,8 +1,12 @@
 import { DocumentQuery, UniqueId } from "documentdb";
 
+export const ENTITY_PROPERTY_NAME = "_entity";
+export const ENTITY_PROPERTY_PARAMETER_NAME = "@" + ENTITY_PROPERTY_NAME;
+export const ENTITY_TYPE_SEPARATOR = "-";
+
 export interface EntityDocument extends UniqueId {
   /** The entity type (e.g. products, users...) */
-  _entity: string;
+  [ENTITY_PROPERTY_NAME]: string;
 }
 
 export enum Direction {
@@ -105,6 +109,13 @@ class DocumentQueryBuilderImpl implements DocumentQueryBuilder {
           case Operator.Gte:
           case Operator.Lt:
           case Operator.Lte:
+            if ((queryParts.parameterName === "id")
+             && this.sqlParameters[ENTITY_PROPERTY_PARAMETER_NAME]
+             && !queryParts.parameterValue.startsWith(this.sqlParameters[ENTITY_PROPERTY_PARAMETER_NAME])) {
+              queryParts.parameterValue =
+                // tslint:disable-next-line:max-line-length
+                `${this.sqlParameters[ENTITY_PROPERTY_PARAMETER_NAME]}${ENTITY_TYPE_SEPARATOR}${queryParts.parameterValue}`;
+            }
             this.where(
               `${queryParts.leftPath} ${queryParts.operator} @${queryParts.parameterName}`,
               { [queryParts.parameterName]: queryParts.parameterValue });
@@ -117,6 +128,11 @@ class DocumentQueryBuilderImpl implements DocumentQueryBuilder {
             const parameterNames = queryParts.parameterValue.map((_, i) => `@${queryParts.parameterName}${i}`);
             this.whereConditions.push(`${queryParts.leftPath} IN (${parameterNames.join(", ")})`);
             queryParts.parameterValue.forEach((x, i) => {
+              if ((queryParts.parameterName === "id")
+              && this.sqlParameters[ENTITY_PROPERTY_PARAMETER_NAME]
+              && !x.startsWith(this.sqlParameters[ENTITY_PROPERTY_PARAMETER_NAME])) {
+                x = `${this.sqlParameters[ENTITY_PROPERTY_PARAMETER_NAME]}${ENTITY_TYPE_SEPARATOR}${x}`;
+              }
               this.sqlParameters[parameterNames[i]] = x;
             });
             break;